backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Codex-as-chat backend service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "openai>=1.52",
    "greenlet>=3.2.4",
]

[project.optional-dependencies]
dev = [
    "black>=24.0",
    "ruff>=0.5",
]

[tool.uvicorn]
factory = "app:create_app"
